Make generic VirtualList component
